Refactor: Add utils/db
